+++ src/main/java/com/project/cheerha/domain/bookmark/controller/BookmarkController.java
@@ -2,16 +2,22 @@ package com.project.cheerha.domain.bookmark.controller;
 
 import com.project.cheerha.common.annotation.Auth;
 import com.project.cheerha.common.dto.AuthUser;
-import com.project.cheerha.domain.bookmark.dto.CreateBookmarkRequestDto;
-import com.project.cheerha.domain.bookmark.dto.DeleteBookmarkRequestDto;
-import com.project.cheerha.domain.bookmark.dto.ReadBookmarkResponseDto;
+import com.project.cheerha.domain.bookmark.dto.request.CreateBookmarkRequestDto;
+import com.project.cheerha.domain.bookmark.dto.request.DeleteBookmarkRequestDto;
+import com.project.cheerha.domain.bookmark.dto.response.ReadBookmarkResponseDto;
 import com.project.cheerha.domain.bookmark.service.BookmarkService;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
-import org.springframework.web.bind.annotation.*;
+import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
+import org.springframework.web.bind.annotation.RestController;
 
 @RestController
 @RequestMapping("/bookmarks")
