+++ src/main/java/com/project/cheerha/domain/bookmark/dto/request/CreateBookmarkRequestDto.java
@@ -1,10 +1,10 @@
-package com.project.cheerha.domain.bookmark.dto;
+package com.project.cheerha.domain.bookmark.dto.request;
 
 import jakarta.validation.constraints.NotNull;
 
 public record CreateBookmarkRequestDto(
-        @NotNull
-        Long jobOpeningId
+    @NotNull
+    Long jobOpeningId
 ) {
 
 }
