+++ src/main/java/com/project/cheerha/domain/bookmark/repository/BookmarkRepository.java
@@ -2,13 +2,13 @@ package com.project.cheerha.domain.bookmark.repository;
 
 import com.project.cheerha.domain.bookmark.entity.Bookmark;
 import com.project.cheerha.domain.jobOpening.entity.JobOpening;
+import java.util.Optional;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-import java.util.Optional;
-
 public interface BookmarkRepository extends JpaRepository<Bookmark, Long> {
+
     // userId와 jobOpeningId로 북마크 조회
     boolean existsByUserIdAndJobOpeningId(Long userId, Long jobOpeningId);
